Resolved the 14-month-stale merge by taking current `master` for all files
and re-applying the PR's LIMIT push-down logic in
`transformQueryForExternalDatabase.cpp`, `transformQueryForExternalDatabaseAnalyzer.cpp`,
and its unit test, adapting to master's `make_intrusive` AST allocation.

+++ src/Common/hasLinuxCapability.cpp
@@ -1,6 +1,6 @@
 #if defined(OS_LINUX)
 
-#include "hasLinuxCapability.h"
+#include <Common/hasLinuxCapability.h>
 
 #include <syscall.h>
 #include <unistd.h>
@@ -8,6 +8,7 @@
 #include <cstdint>
 #include <base/types.h>
 #include <Common/Exception.h>
+#include <Common/ErrnoException.h>
 
 
 namespace DB
@@ -60,7 +61,7 @@ static Capabilities getCapabilities()
 bool hasLinuxCapability(int cap)
 {
     static Capabilities capabilities = getCapabilities();
-    return (1 << cap) & capabilities.effective;
+    return (UInt64(1) << cap) & capabilities.effective;
 }
 
 }
